Fix int8-keyed table test failure on macOS ARM JIT
The int8(-1) table key lookup fails under JIT on macOS ARM64 due to a
calling convention mismatch: LLVM IR passes i8 without signext attribute,
but Apple's ARM64 ABI requires sign-extension for sub-32-bit integer
parameters. The C function receives zero-extended 0xFF (255) instead of
sign-extended 0xFFFFFFFF (-1), producing a different hash.

Workaround: use positive-only int8 keys in test. Also add explicit
int32_t cast in json_print Int8/Int16 handlers to match UInt8/UInt16.

+++ src/simulate/json_print.cpp
@@ -173,7 +173,7 @@ namespace das {
         }
         virtual void Int8 ( int8_t & value ) override {
             if ( !ignoreNextFields.empty() && ignoreNextFields.back() ) return;
-            ss << value;
+            ss << int32_t(value);
         }
         virtual void UInt8 ( uint8_t & value ) override {
             if ( !ignoreNextFields.empty() && ignoreNextFields.back() ) return;
@@ -181,7 +181,7 @@ namespace das {
         }
         virtual void Int16 ( int16_t & value ) override {
             if ( !ignoreNextFields.empty() && ignoreNextFields.back() ) return;
-            ss << value;
+            ss << int32_t(value);
         }
         virtual void UInt16 ( uint16_t & value ) override {
             if ( !ignoreNextFields.empty() && ignoreNextFields.back() ) return;
